+++ lib/modules/state/state_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: state
-// Updated at: 2026-07-30 15:29:56
-// Commit: feat(state): add bi-directional event bus handler for state
+// Updated at: 2026-07-30 15:38:23
+// Commit: perf(state): optimize memory consumption in state list viewport
 
 class stateComponentState {
-  final String commitRef = '3245';
-  final String lastMessage = 'feat(state): add bi-directional event bus handler for state';
-  final DateTime timestamp = DateTime.parse('2026-07-30T15:29:56+05:00');
+  final String commitRef = '3246';
+  final String lastMessage = 'perf(state): optimize memory consumption in state list viewport';
+  final DateTime timestamp = DateTime.parse('2026-07-30T15:38:23+05:00');
 }